Update core.date.php
Changes in detail:
Before calling new \DateTimeZone($timezone), the system checks whether the time zone passed is contained in the list of valid time zones (\DateTimeZone::listIdentifiers()). If not, the time zone is set to 'UTC' to ensure that no error occurs.
If an invalid time zone is passed, a fallback to 'UTC' is performed. This is a safe standard and avoids triggering errors if invalid values occur.

It is now ensured that no infinite loop is created in the callback function for preg_replace_callback() (server error 503) and that invalid or unhandled format characters are properly intercepted.
If an unknown or invalid format character is found, it is either replaced with a default value or left unchanged in the output.